northd: Commit newly initiated flows in DNAT reply direction
On a Gateway router with DNAT flows configured, the router will
translate the destination IP address from (A) to (B). Reply packets
from (B) are correctly UNDNATed in the reverse direction.

However, if a new connection is established from (B), this flow is never
committed to conntrack and, as such, is never established. This will
cause OVS datapath flows to be added that match on the ct.new flag.

For software-only datapaths this is not a problem. However, for
datapaths that offload these flows to hardware, this may be problematic
as some devices are unable to offload flows that match on ct.new.

This patch resolves this by committing these flows in the router
DNAT zone.

This patch also moves the UNDNAT flow to the lr_out_undnat table.

Signed-off-by: Mark Gray <mark.d.gray@redhat.com>